+++ package.json
@@ -20,7 +20,8 @@
     "@mui/material-nextjs": "^5.15.5",
     "next": "14.0.3",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "recoil": "^0.7.7"
   },
   "devDependencies": {
     "@types/node": "^20",
